provide generic peer service for plumtree protocol
In order for the plumtree protocol to operate it must be provided with
a list of peer nodes in order to populate and maintain the sets of eager
and lazy peers. This commits a minimum viable peer service so that
plumtree can be used a library outside of riak_core.

Upon startup each node will create an ETS table called `cluster_state`
which will be populated with a unique actorid and a cluster_state object
that is represented by an ORSWOT borrowed from `riak_dt`. As nodes join
and leave the cluster_state object will be modified to reflect the change
and then optimistically gossipped to other peers.

This current peer service implementation is limited in several ways, which
I plan to fix in future updates:

- missing health/liveness checks against connected peers
- unable to force-remove failed nodes from cluster_state
- cluster_state object is simply a primitive list of nodes
- peer service is hard-coded to plumtree_broadcast, and not pluggable
- failed nodes cannot auto-rejoin clusters
- gossip protocol is random, which is probabilistically acceptable for small
clusters, but the time till quiesence will increase as more nodes join

Further considerations:

I elected to bring in riak_dt as a dependency early, since it made managing
cluster_state much simpler than relying on other data structures and rolling
custom causality logic. In the future I hope to replace plumtree's object
representation with CRDTs, to allow for more flexible data structures and to
escape the LWW semantic. This may also be beneficial to removing/augmenting
the LevelDB backed AAE.