track upstream riak_dt
Support for OTP 17 landed in riak_dt upstream, under the `2.1.0` tag. This
is the only difference between upstream and the existing helium fork.

Tracking upstream from plumtree is advantageous for obvious reasons,
especially considering the existing open PR to support the Causual LWW
Register, which if merged will provide the same semantics as dvvset.erl,
making it that much easier to provide CRDTs as plumtree_metadata_object's
underlying representation.