Force nodes to delete cluster_state when leaving
When a node leaves a cluster it also shutsdown, for practical reasons.

Previously, once a node could confirm it had been removed from a peer's
cluster_state ORSWOT, it would rewrite a new cluster_state containing
only itself prior to shutting down. This caused an edge cause where,
once restarted, the local node would read the cluster_state data from
disk, using a no-longer valid actor id. When attempting to rejoin the
cluster which was left previously, the clock of the remote nodes would
dominate the local node's, resulting in a join where the local node is
connected over disterl, but not reflected in the membership list.

To avoid this case, a leaving node should delete the cluster_state
file which has been persisted to disk.

A case has been added to the cluster_membership_SUITE to test for this
behaviour in the future.